Commit 6
Forgot to mention in last commit that I created AReg and DReg arrays to store the value currently in the register. Going to need this so we know how to set the LEDs coming out of them. Added a function to set LEDs coming out of DReg. Added/fixed function for bottom right and top right repeaters. Also wrote functions for top middle splitter (needed one to decide to send to DReg or not and 1 to decide whether or not to send towards memory/AReg).