Commit 7
Corrected function for top left repeater. Created two functions for bottom left splitter to decide whether or not to send its inputted data to memory, AReg, or both. Splitter to the mem now sets the memory location specified by the A address to the appropriate value (if the splitter specifies that it will go to mem). Same thing with splitter to AReg. Also took care of ARegtoMem LEDs.